\name{tickdatabase-package}
\alias{tickdatabase-package}
\alias{tickdatabase}
\docType{package}
\title{
\packageTitle{tickdatabase}
}
\description{
\packageDescription{tickdatabase}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tickdatabase}
\packageIndices{tickdatabase}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{tickdatabase}

Maintainer: \packageMaintainer{tickdatabase}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
